trader/email-sync/email-ingestion/src/requirements.txt
boto3==1.34.0
botocore==1.34.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
PyJWT==2.8.0
requests==2.31.0 